Consistent and reproducible datasets (#22)
* feat: Prototype for datasets.emissions()

* cfg: Requirement on pooch

Using package pooch to simplify the code to download the source data.

* feat: WebFire factors dataset

* fix: Missing HTTPDownloader

* feat: fetch_scc() - Source Classification Codes

* refact: Using root path

We might not need to isolate the data in groups since a hash is added to
the filenames.

* doc: Adding more information

* fix: Using an sha256 hash

* doc: Info on fetch_emission()

* feat: Download NEI data

Although it's not decompressed automatically due to a a bad formating
from the data providers, these new functions download the zip files to
be manually processed later.

* refact: Use WebFire emission factors from dataset

This guarantee efficiency on re-using already downloaded data and remove
requirements to manage local directories to manage such dataset.

* doc: Improving documentation for fetch_webfirefactors()

* doc: datasets.fetch_emission()

* Getting path to excel file instead

A partial refactoring. For now it just download the raw file and use
legacy code to parse it.

* refact: Using datasets instead of ghgrp_fac_unit.download_unit_data

* style:

* doc: Note for development

* cfg: Some dependency requirements run as it is

Locking on some old versions to be able to access some of the source
datasets. Later I'll work on release those versions.

* doc: Initiating a development document