refact: Downloading data with pooch
Avoiding the custom download procedure in favor of using pooch. The
reasoning here is to simplify FIED package by using well stabilished
open source solutions such as pooch. This also reduce the need of
tracking local directories to hold raw input data.

This is a transition commit to minimize changes in the original code.