Limit WebFires data to simple factors only (#42)
* fix: Limit webfr to simple factors only

There are other types of formula beyond simple factors, such as range,
formula, etc. Those forced the factor column to be read as string to
accomodate all the types. Hence, simple ordering doesn't lead to the
maximum observed factor, resulting in errors.

For instance, '5e-10' is after '1000' when compared as strings.

* doc: Extending comment on ignored records

* doc: Inline comment on what is happening